Refactor: renames extractExceptionsFromAho to generic extractExceptionsFromXml
