+++ packages/core/phase1/parse/parseAhoXml/index.ts
@@ -1,4 +1,4 @@
-export { default as extractExceptionsFromAho } from "./extractExceptionsFromAho"
+export { default as extractExceptionsFromXml } from "./extractExceptionsFromXml"
 export { default as mapXmlCxe01ToAho } from "./mapXmlCxe01ToAho"
 export { default as parseAhoXml } from "./parseAhoXml"
 export { mapXmlToAho } from "./parseAhoXml"
