+++ packages/core/phase1/tests/helpers/processMessage.ts
@@ -4,7 +4,7 @@ import { v4 as uuid } from "uuid"
 import CoreAuditLogger from "../../../lib/CoreAuditLogger"
 import type { AnnotatedHearingOutcome } from "../../../types/AnnotatedHearingOutcome"
 import MockPncGateway from "../../../comparison/lib/MockPncGateway"
-import extractExceptionsFromAho from "../../parse/parseAhoXml/extractExceptionsFromAho"
+import extractExceptionsFromAho from "../../parse/parseAhoXml/extractExceptionsFromXml"
 import parseSpiResult from "../../parse/parseSpiResult"
 import transformSpiToAho from "../../parse/transformSpiToAho"
 import phase1Handler from "../../phase1"
